# backend/openapi/skeleton.yaml
...
openapi: 3.0.3
info:
  title: SALEC API (skeleton)
  description: |
    Infratuzilma va umumiy kontrakt tekshiruvi. Domain marshrutlari bosqichma-bosqich qo‘shiladi.
  version: "0.0.0"
  license:
    name: Unlicense
servers:
  - url: http://127.0.0.1:4000
    description: Local backend
paths:
...
  /api/{slug}/dashboard/stats:
    get:
      summary: Dashboard summary stats
      operationId: dashboardStats
      tags: [Dashboard]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantSlug"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/{slug}/dashboard/supervisor:
    get:
      summary: Supervisor dashboard snapshot
      operationId: dashboardSupervisor
      tags: [Dashboard]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantSlug"
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/{slug}/dashboard/sales:
    get:
      summary: Sales dashboard snapshot
      operationId: dashboardSales
      tags: [Dashboard]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantSlug"
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/{slug}/dashboard/finance:
    get:
      summary: Finance dashboard snapshot
      operationId: dashboardFinance
      tags: [Dashboard]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantSlug"
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /api/{slug}/mobile/agent-config:
    get:
      summary: Mobile agent config
      operationId: mobileAgentConfig
      tags: [Mobile]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantSlug"
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/{slug}/mobile/sync/full:
    post:
      summary: Mobile full sync
      operationId: mobileSyncFull
      tags: [Mobile]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantSlug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/{slug}/mobile/sync/delta:
    post:
      summary: Mobile delta sync (single entity type)
      operationId: mobileSyncDelta
      tags: [Mobile]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantSlug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/{slug}/mobile/orders/pending:
    get:
      summary: Pending offline orders count
      operationId: mobileOrdersPending
      tags: [Mobile]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantSlug"
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/{slug}/mobile/fcm/register:
    post:
      summary: Register FCM device token
      operationId: mobileFcmRegister
      tags: [Mobile]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantSlug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/{slug}/mobile/orders/enqueue:
    post:
      summary: Enqueue offline order
      operationId: mobileEnqueueOrder
      tags: [Mobile]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantSlug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "201":
          description: Created
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantSlug:
      name: slug
      in: path
      required: true
      schema:
        type: string
        example: test1
  schemas:
    ApiError:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: string
          example: NotReady
        requestId:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true